Subtle Changes to the base and working on parts
